requirement.txt
streamlit
rembg
pillow
numpy
onnxruntime


